refactor(client): extract useConnectionType from P2PTransfer
Fourth step of decomposing P2PTransfer.tsx into focused hooks. Move the
direct-vs-relay detection (checkConnectionType reading peer ICE stats, plus its
5s polling interval and the connectionType state) into
client/hooks/useConnectionType.ts.

The hook returns { connectionType, startPolling, stopPolling, reset }; the two
peer connect/close handlers and the mount cleanup now call those instead of
managing the interval inline. Behavior-preserving: the detection logic is moved
verbatim and the connectionType reads (badge, isRelayOverLimit, analytics,
floe-connection-status dispatch) are unchanged.

Deliberately untouched: the sender's separate inline getStats relay-fallback
validation, which is a different concern and not covered by e2e.
